Extract seed content loading from orchestrator
